tlcl_measure() extends a PCR with the SHA1 hash of a memory region.
Add tlcl_measure() to compute the SHA1 hash of a memory region
and extend a given PCR with this hash.  This also links it into
the romstage so that it is possible to measure the bootblock
and any files loaded from the CBFS.

SHA1 implementation copied from the chromeec directory.